test(fuzz): frame a mutated payload with the schema, as wire.ts requires
The robustness classifier splits a mutated payload two ways: well-formed
protobuf read differently is a codec bug, bytes that do not frame are
undefined behaviour. It decided that with `canonicalWire(bytes)` and no
schema, so it framed only the top level — while the real decoders descend
into submessages using the schema. `wire.ts` states the rule it broke:
every caller that has a schema passes it, and this one has `path`.

Passing the schema alone changed nothing, because `scanFrom` fell back to
the raw hex whenever a nested scan failed. That fallback is right without
a schema — the payload could be a string or a `bytes` field that merely
looks like protobuf — and wrong with one: a field the schema declares a
message, whose contents do not frame, makes the record malformed. Both
halves were needed.

`schemaAt` and the field-number recovery it depends on move out of the
codec differential into `harness/schema-context.ts`, which is what lets a
second fuzzer have a schema at all.

One of the four findings reclassifies to proto:mutation-interpretation,
where the entry beside it already calls that undefined. I predicted three
would move; only one did, because my probe treated every length-delimited
field as a submessage where the schema descends only into declared ones —
in the other two the corruption sits in a `bytes` or `string` field,
which is genuinely opaque.

The three that remain frame at the top level and through every declared
submessage. Two decoders reading well-formed protobuf differently is a
parser differential, and no allowlist entry is written for it: a
predicate loose enough to cover it would have to excuse "same key,
different value" on mutated input, which is the property this target
exists to protect.
